refactor: split background to multiple js files
Split background to multiple js files so that they are ready to test
