+++ package.json
@@ -1,5 +1,6 @@
 {
   "name": "chrome_extension_react",
+  "type": "module",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
